Implement returning after levels
We will need to return with data after each level
is completed. This commit adds code to simply
launch a new level after the level is completed or
the game is lost. The points are persisted through
the process. Next we will implement this to only
keep going if the level was actually won.